refactor: move basic component to design system module and custom component to the ui
